Refactor package and class names
