+++ component/publisher-client/pom.xml
@@ -71,14 +71,14 @@
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Export-Package>
-                            !org.wso2.am.analytics.publisher.reporter.choreo.*;version="${project.version}",
+                            !org.wso2.am.analytics.publisher.reporter.cloud.*;version="${project.version}",
                             !org.wso2.am.analytics.publisher.reporter.choreo;version="${project.version}",
                             org.wso2.am.analytics.publisher.client.*;version="${project.version}",
                             org.wso2.am.analytics.publisher.reporter.*;version="${project.version}",
                             org.wso2.am.analytics.publisher.exception.*;version="${project.version}"
                         </Export-Package>
                         <Private-Package>
-                            org.wso2.am.analytics.publisher.reporter.choreo.*;version="${project.version}",
+                            org.wso2.am.analytics.publisher.reporter.cloud.*;version="${project.version}",
                         </Private-Package>
                         <Import-Package>
                             com.google.gson.*;version="${gson.import.version.range}",
