+++ component/publisher-client/src/main/java/org/wso2/am/analytics/publisher/reporter/cloud/QueueWorker.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.am.analytics.publisher.reporter.choreo;
+package org.wso2.am.analytics.publisher.reporter.cloud;
 
 import com.google.gson.Gson;
 import org.apache.log4j.Logger;
