+++ findbugs-exclude.xml
@@ -19,7 +19,7 @@
 <FindBugsFilter>
     <!--api-->
     <Match>
-        <Class name="org.wso2.am.analytics.publisher.reporter.choreo.EventQueue"/>
+        <Class name="org.wso2.am.analytics.publisher.reporter.cloud.EventQueue"/>
         <Bug pattern="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE"/>
     </Match>
 </FindBugsFilter>
